+++ config/routes.rb
@@ -1,13 +1,14 @@
 Rails.application.routes.draw do
 
-  resources :orders, only:[:new, :create, :index, :show] do
+  resources :orders, controller: 'public/orders', only:[:new, :create, :index, :show] do
     member do
       post 'confirm'
       get 'thanks'
     end
   end
   
-  resources :addresses, only:[:index, :create, :edit, :update, :destroy]
+  resources :addresses, only: [:index, :create, :edit, :update, :destroy]
+
 
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
