+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2025_02_16_080947) do
+ActiveRecord::Schema.define(version: 2025_02_16_133102) do
 
   create_table "addresses", force: :cascade do |t|
     t.integer "customer_id", null: false
@@ -20,6 +20,7 @@ ActiveRecord::Schema.define(version: 2025_02_16_080947) do
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["customer_id"], name: "index_addresses_on_customer_id"
+  end
 
   create_table "admins", force: :cascade do |t|
     t.string "email", default: "", null: false
